Split comparison unit tests to files per class
